YaaS/yaas_server/requirements.txt
asgiref==3.8.1
boto3==1.38.8
botocore==1.38.8
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
Django==5.2
django-appconf==1.1.0
django-browser-reload==1.18.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-storages==1.14.6
django-tailwind==4.0.1
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud==0.34.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.4.2
numpy==2.2.5
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rjsmin==1.2.2
rsa==4.9.1
s3transfer==0.12.0
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
